Structured doctors routes
